requirements.txt
nougat-ocr==0.1.17
pdf2image==1.17.0
langchain-ollama==0.2.2
langchain-core==0.3.27
transformers==4.47.1
torch==2.1.2
torchvision==0.16.2
Pillow==11.0.0
pypdf==5.1.0
pypdfium2==4.30.1
sentencepiece==0.2.0
nltk==3.9.1
python-Levenshtein==0.26.1
albumentations==1.4.23
opencv-python-headless==4.10.0.84
datasets[vision]==3.2.0
lightning==2.4.0
sconf==0.2.5 